refactor : 패키지 변경
- Parser 에서 converter 를 사용하는 것으로 변경함에 따라 Converter 패키지를 parser 내부로 이동
