--- src/main/java/calculator/converter/IntegerStringConverter.java
@@ -1,8 +1,0 @@
-package calculator.converter;
-
-public class IntegerStringConverter implements calculator.converter.StringConverter<Integer> {
-
-    public Integer convert(String numb) {
-        return Integer.parseInt(numb);
-    }
-}
