+++ src/main/java/calculator/parser/converter/IntegerStringConverter.java
@@ -1,0 +1,8 @@
+package calculator.parser.converter;
+
+public class IntegerStringConverter implements StringConverter<Integer> {
+
+    public Integer convert(String numb) {
+        return Integer.parseInt(numb);
+    }
+}
