+++ src/main/java/calculator/parser/converter/StringConverter.java
@@ -1,4 +1,4 @@
-package calculator.converter;
+package calculator.parser.converter;
 
 public interface StringConverter<T> {
     T convert(String input);
