+++ src/test/java/calculator/converter/IntegerStringConverterTest.java
@@ -1,5 +1,7 @@
 package calculator.converter;
 
+import calculator.parser.converter.IntegerStringConverter;
+import calculator.parser.converter.StringConverter;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
